Task 4. Retrieve employees in Marketing
Your team is updating employee machines, and you need to obtain the information about employees in the 'Marketing' department who are located in all offices in the East building (such as 'East-170' or 'East-320').

Write a SQL query to retrieve this information from the employees table. Select all columns and include filters on the department and office columns to return only the needed records.